Vehicle sales statistics by country and type

This data frame contains annual data on vehicle sales statistics by country and type between the years 2005 and 2021 (inclusive) from the OICA. Standard country / area codes come from the United Nations Statistics Division. The table below shows a detailed description of each variable.

data(sales_country)

Format

Variable	Description

var	Description

Source

Raw data downloaded from the Organisation Internationale des Constructeurs d'Automobiles (OICA) and the United Nations Statistics Division

Examples

data(sales_country)

head(sales_country)
#> # A tibble: 6 × 16
#>    year country type      n region subregion     inter…¹ least…² land_…³ small…⁴
#>   <dbl> <chr>   <chr> <dbl> <chr>  <chr>         <chr>     <dbl>   <dbl>   <dbl>
#> 1  2005 Albania pc      800 Europe Southern Eur… NA            0       0       0
#> 2  2006 Albania pc      800 Europe Southern Eur… NA            0       0       0
#> 3  2007 Albania pc     1600 Europe Southern Eur… NA            0       0       0
#> 4  2008 Albania pc     1600 Europe Southern Eur… NA            0       0       0
#> 5  2009 Albania pc     1000 Europe Southern Eur… NA            0       0       0
#> 6  2010 Albania pc     1600 Europe Southern Eur… NA            0       0       0
#> # … with 6 more variables: code_region <dbl>, code_subregion <dbl>,
#> #   code_intermediate_region <dbl>, code_m49 <dbl>, code_iso_alpha2 <chr>,
#> #   code_iso_alpha3 <chr>, and abbreviated variable names ¹​intermediate_region,
#> #   ²​least_developed, ³​land_locked_developing, ⁴​small_island_developing